+++ test/dummy/config/application.rb
@@ -2,6 +2,7 @@
 
 require "rails"
 require "active_record/railtie"
+require "active_job/railtie"
 require "action_controller/railtie"
 require "solid_queue"