+++ test/dummy/config/routes.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
+  get "answer", to: "requests#answer"
   match "fail", to: "requests#failure", via: [ :get, :post ]
 end